App/backend-v2/sentiment-analysis/requirements.txt
colorlog==6.8.2
fastapi==0.110.0
pydantic==2.6.4
pydantic-core==2.16.3
httpx==0.27.0
asyncio==3.4.3
uvicorn==0.29.0
python-dotenv==1.0.1
nltk==3.8.1
transformers==4.41.2
torch==2.3.1